+++ Missionframework/ui/mission_params.hpp
@@ -251,6 +251,12 @@ class Params {
         texts[] = {$STR_PARAMS_DISABLED, $STR_PARAMS_ENABLED};
         default = 1;
     };
+    class AttackedSectorRespawn {
+        title = $STR_PARAMS_ATTACKEDSECTORRESPAWN;
+        values[] = {0, 1};
+        texts[] = {$STR_PARAMS_DISABLED, $STR_PARAMS_ENABLED};
+        default = 0;
+    };
     class AiLogistics {
         title = $STR_PARAMS_AILOGISTICS;
         values[] = {0, 1};