+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added: ACE, TFAR and ACRE items to GM West and East arsenal presets.
 * Added: Config value for radio tower classnames.
 * Added: Parameter to disable Zeus for commander.
+* Added: Disable respawn on attacked FOBs (adjustable via parameter)
 * Removed: T-14 from RHS AFRF preset.
 * Tweaked: "FOB " removed from resource overlay, so it's just e.g. "ALPHA" again.
 * Fixed: Sector monitor got stuck after sector cap was reached until restarting the server.